Create Hardware Generation Tool
The project involved Register Transfer Level (RTL) Design, simulation, synthesis and testbench development for verification of a configurable Artificial Neural Network in SystemVerilog and then creating a parameterizable HDL Generation tool in C++ to create an optimized 3-Layer Neural Network. Each Layer of this Network is individually configurable for Size and Parallelism and hence we can optimize the performance based on the available resources. The generator has an Optimization algorithm which decides the Parallelism for each layer based on Multiplier Budget allocated to maximize Performance.